Symbolic tags for simplifier phases
Every simplifier phase can have an arbitrary number of tags and multiple
phases can share the same tags. The tags can be used as arguments to
-ddump-simpl-phases to specify which phases are to be dumped.
For instance, -ddump-simpl-phases=main will dump the output of phases 2, 1 and
0 of the initial simplifier run (they all share the "main" tag) while
-ddump-simpl-phases=main:0 will dump only the output of phase 0 of that run.

At the moment, the supported tags are:

  main                 The main, staged simplifier run (before strictness)
  post-worker-wrapper  After the w/w split
  post-liberate-case   After LiberateCase
  final                Final clean-up run

The names are somewhat arbitrary and will change in the future.