Print out rational literals correctly in External Core
The External Core printer was printing out rational literals of the
form: 
2.0e-2
when the External Core grammar doesn't allow this. (This
bug has apparently been there since the beginning...)

It's now printing rationals in the same form that (show (r::Rational))
does. This requires a parser change as well (soon to come.)