Fix External Core interpreter
The External Core interpreter works (in a limited sense).
For details, see the README.

This means we now have a marginally functioning set of
External Core tools.

The other exciting change is that the test driver (Driver.hs)
now computes module dependencies automatically instead of
having a wired-in list of library modules.