Refactoring in H.Interface.Create
We were creating a doc map, a declaration map and a list of entities
separately by going through the HsGroup. These structures were all used
to build the interface of a module.

Instead of doing this, we can start by creating a list of declarations
from the HsGroup, then collect the docs directly from this list 
(instead of using the list of entities), creating a documentation map.

We no longer need the Entity data type, and we can store a single
map from names to declarations and docs in the interface, instead of
the declaration map and the doc map.

This way, there is only one place where we filter out the declarations
that we don't want, and we can remove a lot of code.

Another advantage of this is that we can create the exports directly
out of the list of declarations when we export the full module contents.
(Previously we did a look up for each name to find the declarations).
This is faster and removes another point where we depend on names to
identify exported declarations, which is good because it eliminates
problems with instances (which don't have names).