Preparation for rendering instances as separate declarations
We want to be able to render instances as separate declarations. So we remove
the Name argument of ExportDecl, since instances are nameless.

This patch also contains the first steps needed to gather type family instances 
and display them in the backend, but the implementation is far from complete.
Because of this, we don't actually show the instances yet.