Be compatible with GHC 6.8.3
The cabal file is converted to use the "new" syntax with explicit Library
and Executable sections.

We define the __GHC_PATCHLEVEL__ symbol using a conditinal cpp-options field
in the cabal file. (Ideally, Cabal would define the symbol for us, like it does
for __GLASGOW_HASKELL__).

We use these symbols to #ifdef around a small difference between 6.8.2 and 6.8.3.
    
Previously, we only supported GHC 6.8.2 officially but the dependencies field
said "ghc <= 6.9". This was just for convenience when testing against the (then 
compatible) HEAD version of GHC, and was left in the release by mistake.
  
Now, we support both GHC 6.8.2 and 6.8.3 and the dependencies field
correctly reflects this.