Make comparison on equalities work right (ie look at the types)
This patch makes
	(s1~t1) == (s2~t2) iff s1==s2, t1==t2

The comparison and equality instances for Insts are really only
used in making the AvailEnv in TcSimplify, which equalities are
never put into, which is why we've gotten away with this for so long.

But I'm now using 'nub' in the error messages for equalities,
and in any case it makes sense to have the right equality!