Retain unfoldings even with SimplGently
When binding x = e, we now attach an unfolding to 'x' even if
it won't be used because SimplGently is on. 

Reason: the specialiser runs right after SimplGently, and it (now)
only gathers call information for calls whose dictionary arguments are
"interesting" -- i.e. have an unfolding of some kind.