Improve float-in somewhat
See Note [Floating in past a lambda group]. The new thing
here is that we previously were not floating in past a group
that had (\(a:*) \(x:State# a). e), with a type var *and* a
one-shot lambda.  

The fix makes wave4main allocate 2% less, and doesn't change any other
nofib number, so it's not a big deal.