Re-adjust interaction between -ddump flags and force-recompilation
If you say -ddump-xx we effectively add -fforce-recomp, so that you
see your dump output.  But this works badly in --make mode, because
you get the ddump output for every module, which is probably not what
you want.  This patch forces recompilation with -ddump-X only in one-shot
mode.  

Of course, this only affects people using -ddump options.