Add extra WARN test
This warning tests that the arity of a function does not decrease.
And that it's at least as great as the strictness signature.

Failing this test isn't a disater, but it's distinctly odd and 
usually indicates that not enough information is getting propagated
around, and hence you may get more simplifier iterations.